[修复] 系统界面-通知与控制中心-GMS服务添加磁贴失败的问题
