+++ app/src/main/java/com/sevtinge/cemiuiler/module/SystemUI.java
@@ -19,13 +19,14 @@ import com.sevtinge.cemiuiler.module.systemui.controlcenter.AddBlurEffectToNotif
 import com.sevtinge.cemiuiler.module.systemui.controlcenter.CCGrid;
 import com.sevtinge.cemiuiler.module.systemui.controlcenter.CompactNotificationsHook;
 import com.sevtinge.cemiuiler.module.systemui.controlcenter.FixMediaControlPanel;
+import com.sevtinge.cemiuiler.module.systemui.controlcenter.GmsTile;
 import com.sevtinge.cemiuiler.module.systemui.controlcenter.MuteVisibleNotifications;
 import com.sevtinge.cemiuiler.module.systemui.controlcenter.NotificationWeather;
 import com.sevtinge.cemiuiler.module.systemui.controlcenter.NotificationWeatherNew;
 import com.sevtinge.cemiuiler.module.systemui.controlcenter.NotificationWeatherOld;
 import com.sevtinge.cemiuiler.module.systemui.controlcenter.QQSGrid;
 import com.sevtinge.cemiuiler.module.systemui.controlcenter.QSControlDetailBackgroundAlpha;
-import com.sevtinge.cemiuiler.module.systemui.controlcenter.FiveAndGmsTile;
+import com.sevtinge.cemiuiler.module.systemui.controlcenter.FiveGTile;
 import com.sevtinge.cemiuiler.module.systemui.controlcenter.QSGrid;
 import com.sevtinge.cemiuiler.module.systemui.controlcenter.QSGridLabels;
 import com.sevtinge.cemiuiler.module.systemui.controlcenter.SmartHome;
@@ -177,8 +178,8 @@ public class SystemUI extends BaseModule {
         // 控制中心
         initHook(new SmartHome(), false);
         initHook(new QSDetailBackGround(), mPrefsMap.getInt("system_control_center_qs_detail_bg", 0) > 0);
-        initHook(new FiveAndGmsTile(), mPrefsMap.getBoolean("security_center_gms_open"));
-        initHook(new FiveAndGmsTile(), mPrefsMap.getBoolean("system_control_center_5g_tile"));
+        initHook(new GmsTile(), mPrefsMap.getBoolean("security_center_gms_open"));
+        initHook(new FiveGTile(), mPrefsMap.getBoolean("system_control_center_5g_tile"));
         initHook(new QSGridLabels(), mPrefsMap.getInt("system_control_center_old_qs_rows", 1) > 1 ||
             mPrefsMap.getBoolean("system_control_center_qs_tile_label"));
         initHook(new MuteVisibleNotifications(), mPrefsMap.getBoolean("system_ui_control_center_mute_visible_notice"));
